requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
colorama==0.4.6
crispy-bootstrap4==2024.1
distro==1.9.0
Django==5.0.3
django-crispy-forms==2.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.7.0
openai==1.54.2
pydantic==2.9.2
pydantic_core==2.23.4
pyngrok==7.2.1
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.1
Flask==2.0.1
requests==2.26.0
Pillow==8.3.1
